requirements.txt
anyio==3.6.1
asgiref==3.5.2
astroid==2.12.5
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
autopep8==1.7.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==37.0.4
databases==0.6.1
Deprecated==1.2.13
dill==0.3.5.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.81.0
flake8==5.0.4
h11==0.13.0
httptools==0.4.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
limits==1.6
MarkupSafe==2.1.1
mccabe==0.7.0
mock==4.0.3
orjson==3.8.0
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.9.2
pyflakes==2.5.0
pylint==2.15.0
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
redis==4.3.4
requests==2.28.1
rsa==4.9
six==1.16.0
slowapi==0.1.6
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.12
uvicorn==0.18.3
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
wrapt==1.14.1
